# competition-app/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Swagger Docs - Adam Harvey",
        "description": "This is the Swagger documentation for my Code Submission for the IS-24 Full Stack Developer Position Code Challenge.",
        "contact": {
            "name": "Adam Harvey",
            "email": "adamgrharvey@gmail.com",
            "url": "https://www.linkedin.com/in/adamgrharvey/"
        },
        "version": "1.0.11"
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "operationId": "getProducts",
                "responses": {
                    "200": {
                        "description": "Responds with object of Product objects.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/Products"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create new product",
                "operationId": "createProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "productName": {
                                        "type": "string",
                                        "description": "The name of the product.",
                                        "example": "NovaForge"
                                    },
                                    "scrumMasterName": {
                                        "type": "string",
                                        "description": "The name of the Scrum Master for the product.",
                                        "example": "Jennifer Garcia"
                                    },
                                    "productOwnerName": {
                                        "type": "string",
                                        "description": "The name of the Product Owner for the product.",
                                        "example": "Jenna Martin"
                                    },
                                    "developers": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "Jenna Martin"
                                        },
                                        "description": "The list of developers who work on the product."
                                    },
                                    "startDate": {
                                        "type": "string",
                                        "description": "The date the project was started.",
                                        "example": "2023-03-27"
                                    },
                                    "methodology": {
                                        "type": "string",
                                        "enum": [
                                            "Waterfall",
                                            "Agile"
                                        ],
                                        "description": "The methodology used to manage the project."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Responds with object of Product objects.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/search/developer/{developerName}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Search for products by developer name",
                "operationId": "searchProductsByDeveloperName",
                "parameters": [
                    {
                        "name": "developerName",
                        "in": "path",
                        "required": true,
                        "description": "Developer name",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns an object containing all products where the developers array contains {developerName}.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/Products"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/search/scrummaster/{scrumMasterName}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Search for products by Scrum Master name",
                "operationId": "searchProductsByScrumMasterName",
                "parameters": [
                    {
                        "name": "scrumMasterName",
                        "in": "path",
                        "required": true,
                        "description": "Scrum Master name",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns an object containing all products where the scrum master name is {scrumMasterName}.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/Products"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{productId}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a product by id",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "description": "Product ID",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product ID not found."
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Edit a product by id",
                "operationId": "editProductById",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "description": "Product ID",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "productName": {
                                        "type": "string",
                                        "description": "The name of the product.",
                                        "example": "NovaForge"
                                    },
                                    "scrumMasterName": {
                                        "type": "string",
                                        "description": "The name of the Scrum Master for the product.",
                                        "example": "Jennifer Garcia"
                                    },
                                    "productOwnerName": {
                                        "type": "string",
                                        "description": "The name of the Product Owner for the product.",
                                        "example": "Jenna Martin"
                                    },
                                    "developers": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "Jenna Martin"
                                        },
                                        "description": "The list of developers who work on the product."
                                    },
                                    "startDate": {
                                        "type": "string",
                                        "description": "The date the project was started.",
                                        "example": "2023-03-27"
                                    },
                                    "methodology": {
                                        "type": "string",
                                        "enum": [
                                            "Waterfall",
                                            "Agile"
                                        ],
                                        "description": "The methodology used to manage the project."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "productId": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "productName": {
                                            "type": "string",
                                            "example": "NovaForge"
                                        },
                                        "productOwnerName": {
                                            "type": "string",
                                            "example": "Jenna Martin"
                                        },
                                        "developers": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "a",
                                                "b"
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product ID not found."
                    },
                    "417": {
                        "description": "Expected data format not followed."
                    }
                }
            }
        },
        "/api/health": {
            "get": {
                "tags": [
                    "General"
                ],
                "summary": "Health Check",
                "operationId": "healthCheck",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "Product": {
            "type": "object",
            "properties": {
                "productId": {
                    "type": "integer",
                    "description": "The unique identifier for the product.",
                    "minimum": 1
                },
                "productName": {
                    "type": "string",
                    "description": "The name of the product.",
                    "example": "NovaForge"
                },
                "scrumMasterName": {
                    "type": "string",
                    "description": "The name of the Scrum Master for the product.",
                    "example": "Jennifer Garcia"
                },
                "productOwnerName": {
                    "type": "string",
                    "description": "The name of the Product Owner for the product.",
                    "example": "Jenna Martin"
                },
                "developers": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "example": "Jenna Martin"
                    },
                    "description": "The list of developers who work on the product."
                },
                "startDate": {
                    "type": "string",
                    "description": "The date the project was started.",
                    "example": "2023-03-27"
                },
                "methodology": {
                    "type": "string",
                    "enum": [
                        "Waterfall",
                        "Agile"
                    ],
                    "description": "The methodology used to manage the project."
                }
            },
            "xml": {
                "name": "Order"
            }
        },
        "EditProduct": {
            "type": "object",
            "properties": {
                "productName": {
                    "type": "string",
                    "description": "The name of the product.",
                    "example": "NovaForge"
                },
                "scrumMasterName": {
                    "type": "string",
                    "description": "The name of the Scrum Master for the product.",
                    "example": "Jennifer Garcia"
                },
                "productOwnerName": {
                    "type": "string",
                    "description": "The name of the Product Owner for the product.",
                    "example": "Jenna Martin"
                },
                "developers": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "example": "Jenna Martin"
                    },
                    "description": "The list of developers who work on the product."
                },
                "startDate": {
                    "type": "string",
                    "description": "The date the project was started.",
                    "example": "2023-03-27"
                },
                "methodology": {
                    "type": "string",
                    "enum": [
                        "Waterfall",
                        "Agile"
                    ],
                    "description": "The methodology used to manage the project."
                }
            },
            "xml": {
                "name": "Order"
            }
        },
        "Products": {
            "type": "object",
            "properties": {
                "1": {
                    "type": "object",
                    "properties": {
                        "productId": {
                            "type": "integer",
                            "description": "The unique identifier for the product.",
                            "minimum": 1
                        },
                        "productName": {
                            "type": "string",
                            "description": "The name of the product.",
                            "example": "NovaForge"
                        },
                        "scrumMasterName": {
                            "type": "string",
                            "description": "The name of the Scrum Master for the product.",
                            "example": "Jennifer Garcia"
                        },
                        "productOwnerName": {
                            "type": "string",
                            "description": "The name of the Product Owner for the product.",
                            "example": "Jenna Martin"
                        },
                        "developers": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": [
                                    "Jenna Martin",
                                    "Jared Thomas"
                                ]
                            },
                            "description": "The list of developers who work on the product."
                        },
                        "methodology": {
                            "type": "string",
                            "example": "Agile",
                            "description": "The methodology used to manage the project."
                        }
                    }
                },
                "2": {
                    "type": "object",
                    "properties": {
                        "productId": {
                            "type": "integer",
                            "description": "The unique identifier for the product.",
                            "minimum": 2
                        },
                        "productName": {
                            "type": "string",
                            "description": "The name of the product.",
                            "example": "VaporQuest"
                        },
                        "scrumMasterName": {
                            "type": "string",
                            "description": "The name of the Scrum Master for the product.",
                            "example": "Jennifer Garcia"
                        },
                        "productOwnerName": {
                            "type": "string",
                            "description": "The name of the Product Owner for the product.",
                            "example": "Jared Thomas"
                        },
                        "developers": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": [
                                    "Jared Thomas",
                                    "Trevor Smith",
                                    "Gino Brown",
                                    "Jessica Miller",
                                    "Jenna Martin"
                                ]
                            },
                            "description": "The list of developers who work on the product."
                        },
                        "methodology": {
                            "type": "string",
                            "example": "Waterfall",
                            "description": "The methodology used to manage the project."
                        }
                    }
                }
            },
            "xml": {
                "name": "Order"
            }
        }
    }
}